+++ src/components/buttons/checkbox/Checkbox.style.tsx
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
+import { CheckboxProps } from './Checkbox';
 
-export const Wrapper = styled.button`
+export const Wrapper = styled.button<CheckboxProps>`
   all: unset;
-  cursor: pointer;
+  cursor: ${({ variant }) => (variant === 'teritary' ? 'auto' : 'pointer')};
 `;
