+++ src/components/buttons/checkbox/Checkbox.tsx
@@ -3,7 +3,7 @@
 import * as S from './Checkbox.style';
 
 export type CheckboxProps = {
-  variant: 'primary' | 'secondary';
+  variant: 'primary' | 'secondary' | 'teritary';
   onClick?: () => void;
   isActive?: boolean;
 } & React.ComponentProps<'button'>;
@@ -62,13 +62,28 @@ const switchCheckboxIcon = ({
           />
         </svg>
       );
+    case 'teritary':
+      return (
+        <svg
+          width={18}
+          height={18}
+          viewBox="0 0 18 18"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <path
+            d="M8.99996 0.666687C4.39996 0.666687 0.666626 4.40002 0.666626 9.00002C0.666626 13.6 4.39996 17.3334 8.99996 17.3334C13.6 17.3334 17.3333 13.6 17.3333 9.00002C17.3333 4.40002 13.6 0.666687 8.99996 0.666687ZM7.33329 13.1667L3.16663 9.00002L4.34163 7.82502L7.33329 10.8084L13.6583 4.48335L14.8333 5.66669L7.33329 13.1667Z"
+            fill={isActive ? '#2E74FF' : '#BEC4CD'}
+          />
+        </svg>
+      );
     default:
       return;
   }
 };
 const Checkbox = ({ variant, onClick, isActive, ...props }: CheckboxProps) => {
   return (
-    <S.Wrapper {...props}>
+    <S.Wrapper variant={variant} {...props}>
       {switchCheckboxIcon({ variant, isActive })}
     </S.Wrapper>
   );
