+++ src/components/molecules/resultBox/ResultBox.style.tsx
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 
 export const Wrapper = styled.div`
+  width: 100%;
   padding: 20px 24px;
   display: flex;
   flex-direction: column;
