+++ src/pages/groupLeader/FirstStep/SecondPage/GroupLeaderPage/GroupLeaderPage.style.tsx
