+++ src/pages/groupLeader/SecondStep/SixthPage.tsx
